celery_project/requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
kombu==4.0.2
MangoEngine==1.0.0rc2
mongoengine==0.11.0
olefile==0.44
Pillow==4.0.0
pymongo==3.4.0
pytz==2016.10
redis==2.10.5
requests==2.13.0
selenium==3.0.2
six==1.10.0
vine==1.1.3
